// lib/types/property-type.d.ts
type PropertyType = {
    id: string;
    title: string;
    name: string;
    description: string;
    bedrooms: number;
    bathrooms: number;
    buildYear: string;
    agentId: string;
    slug: string;
    price: string;
    completionStatus: string;
    developerId: string;
    communityId: string;
    community: CommunityType;
    cityId: string;
    offeringTypeId: string;
    offeringType: OfferingType;
    type:UnitType;
    city: CityType;
    images: PropertyImagesType[];
    propertyTypeId: string;
    propertyType:ListingType;
    longitude: string;
    latitude: string;
    cheques: string;
    floor: string;
    permitNumber: string;
    plotSize: number;
    availabilityDate: string;
    parking: string;
    furnished: string;
    referenceNumber: string;
    subCommunity: SubCommunityType;
    serviceCharge: string;
    agent: AgentType;
    size: number;
    updatedAt: string;
    createdAt: string;
}